Revert file renaming
